need to change cmakelists

+++ projects/kernel/code/util/include/util/log.h
@@ -12,8 +12,8 @@ extern "C" {
 #define FLO_NEWLINE 0x01
 
 typedef struct {
-    uint8_t *buffer;
-    uint32_t size;
+    uint32_t *buffer;
+    uint64_t size;
     uint32_t width;
     uint32_t height;
     uint32_t scanline;
